test(quality): resolve CodeFactor findings
