+++ tests/converters/test_openms_consensus.py
@@ -991,142 +991,3 @@ def test_openms_consensus_identity_retains_all_spectrum_references(tmp_path, str
     assert table.column("scan").to_pylist() == [[42, 43]]
     assert table.column("consensus_rt").to_pylist() == pytest.approx([100.123456])
     assert table.schema.metadata[b"identity_composite"] == (b"peptidoform,charge,run_file_name,rt,scan,observed_mz,consensus_rt")
-
-
-@pytest.mark.parametrize("streaming", [False, True])
-def test_openms_consensus_confidence_is_attributed_per_run(tmp_path, streaming):
-    """Each run receives confidence from its own PeptideIdentification."""
-    import pyarrow.parquet as pq
-
-    cx = tmp_path / "multirun-confidence.consensusXML"
-    _write_multirun_confidence_consensusxml(cx)
-    written = OpenMSConsensusConverter().convert(
-        str(cx),
-        str(tmp_path / ("stream" if streaming else "mem")),
-        output_prefix="t",
-        structures=("feature",),
-        streaming=streaming,
-    )
-
-    records = {record["run_file_name"]: record for record in pq.read_table(written["feature"]).to_pylist()}
-    assert records["run_01"]["posterior_error_probability"] == pytest.approx(0.001)
-    assert records["run_01"]["peptide_qvalue"] == pytest.approx(0.001)
-    assert records["run_02"]["posterior_error_probability"] == pytest.approx(0.02)
-    assert records["run_02"]["peptide_qvalue"] == pytest.approx(0.02)
-
-
-class _ScoredHit:
-    """Minimal PeptideHit stand-in carrying a score and optional PEP meta value."""
-
-    def __init__(self, score=None, meta=None):
-        self._score = score
-        self._meta = meta or {}
-
-    def getScore(self):
-        return self._score
-
-    def metaValueExists(self, key):
-        return key in self._meta
-
-    def getMetaValue(self, key):
-        return self._meta[key]
-
-
-def test_pep_of_reads_openms_meta_keys():
-    """PEP is read from whichever meta key OpenMS used."""
-    from qpx.converters.openms_consensus.feature_adapter import pep_of
-
-    assert pep_of(_ScoredHit(meta={"Posterior Error Probability_score": 0.004})) == 0.004
-    assert pep_of(_ScoredHit(meta={"PEP": 0.02})) == 0.02
-    assert pep_of(_ScoredHit(meta={"pep": 0.03})) == 0.03
-    assert pep_of(_ScoredHit()) is None
-
-
-def test_qvalue_of_only_accepts_qvalue_score_types():
-    """A raw search score must never be written to peptide_qvalue.
-
-    Once it is in the column a consumer cannot tell an FDR from a Percolator SVM
-    score, so anything that is not declared a q-value has to stay null
-    (bigbio/qpx#284).
-    """
-    from qpx.converters.openms_consensus.feature_adapter import qvalue_of
-
-    hit = _ScoredHit(score=0.008)
-    for score_type in ("q-value", "Q-Value", "qvalue", "FDR"):
-        assert qvalue_of(hit, score_type) == 0.008, score_type
-    for score_type in ("Posterior Error Probability", "expect", "svm_score", "", None):
-        assert qvalue_of(hit, score_type) is None, score_type
-
-
-def test_qvalue_of_returns_none_without_a_score():
-    from qpx.converters.openms_consensus.feature_adapter import qvalue_of
-
-    assert qvalue_of(_ScoredHit(score=None), "q-value") is None
-
-
-def test_mass_error_ppm_is_computed_from_the_two_mz_values():
-    """mass_error_ppm was null in every artifact although both inputs were present."""
-    from qpx.converters.openms_consensus.feature_adapter import mass_error_ppm
-
-    assert mass_error_ppm(456.5589294433594, 456.5606994628906) == pytest.approx(3.876, abs=1e-2)
-    assert mass_error_ppm(500.0, 499.995) == pytest.approx(-10.0, abs=1e-6)
-
-
-def test_mass_error_ppm_is_none_only_when_not_measurable():
-    """None means absent or unmeasurable — never a real zero.
-
-    An exact match is a genuine 0.0 ppm here: the OpenMS adapters are the only
-    callers and their observed_mz is a measurement, so suppressing it would throw
-    away a real result.
-    """
-    from qpx.converters.openms_consensus.feature_adapter import mass_error_ppm
-
-    assert mass_error_ppm(456.55, 456.55) == 0.0
-    assert mass_error_ppm(456.55, None) is None
-    assert mass_error_ppm(None, 456.55) is None
-    assert mass_error_ppm(0.0, 456.55) is None
-
-
-def test_mass_error_ppm_rejects_non_positive_mz():
-    """A zero m/z is missing data, not a measurement.
-
-    feature_records_for_cf substitutes 0.0 when the ConsensusFeature carries no
-    m/z, so without this guard an unmeasured feature would report roughly
-    -1e6 ppm instead of null.
-    """
-    from qpx.converters.openms_consensus.feature_adapter import mass_error_ppm
-
-    assert mass_error_ppm(456.55, 0.0) is None
-    assert mass_error_ppm(0.0, 456.55) is None
-    assert mass_error_ppm(456.55, -1.0) is None
-    assert mass_error_ppm(-456.55, 456.55) is None
-
-
-def test_unique_is_unknown_without_a_resolved_group(tmp_path):
-    """No resolved group means unknown, not unique.
-
-    A lone peptide evidence is not proof that the peptide maps to one protein, so
-    an unresolved group must leave `unique` null rather than claiming True.
-    Exercised by passing an empty group_map: anchor_protein still resolves from
-    the peptide evidence, but no membership is known.
-    """
-    from qpx.converters.openms_consensus.feature_adapter import (
-        feature_map_info,
-        feature_records_for_cf,
-        load_consensus_map,
-    )
-
-    path = tmp_path / "in.consensusXML"
-    path.write_text(_TMT_CONSENSUSXML)
-    cm = load_consensus_map(str(path))
-    map_info = feature_map_info(cm)
-    cf = list(cm)[0]
-
-    unresolved = feature_records_for_cf(cf, map_info, group_map={})
-    assert unresolved, "expected at least one feature record"
-    for record in unresolved:
-        assert record["anchor_protein"] == "P12345", "anchor still resolves from peptide evidence"
-        assert record["unique"] is None, "unique must be unknown when no group resolved"
-
-    resolved = feature_records_for_cf(cf, map_info, group_map={"P12345": ["P12345"]})
-    assert all(record["unique"] is True for record in resolved), "a resolved group of one is unique"
